requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
aiosqlite==0.17.0
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
Babel==2.14.0
bcrypt==4.1.2
billiard==4.2.0
boto==2.49.0
boto3==1.34.68
botocore==1.34.68
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-admin==1.0.4
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
filetype==1.2.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httptools==0.6.1
humanize==4.9.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.6
iso8601==1.1.0
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.5
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
passlib==1.7.4
pendulum==3.0.0
priority==2.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pwdlib==0.2.0
py3dns==4.0.1
pycparser==2.21
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
PyMySQL==1.1.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
SQLAlchemy-Utils==0.41.1
starlette==0.36.3
starlette-admin==0.13.2
taskgroup==0.0.0a4
time-machine==2.14.1
tomli==2.0.1
tornado==6.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
validate-email==1.3
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wsproto==1.2.0
